fix: delete the file i created and worked on the exsisting file
