+++ apps/vitacare-frontend/src/app/page.tsx
@@ -1,5 +1,5 @@
 import Hero from "@/components/organisms/HeroSection/Hero";
-import KeyFeaturesSection from "@/components/organisms/KeyFeaturesSection/KeyFeaturesSection";
+import KeyFeaturesSection from "@/components/organisms/FeatureGrid/FeatureGrid";
 
 export default function Home() {
   return (
